:recycle: Extract puller.go from job_subscription.go
