+++ job_subscription.go
@@ -8,42 +8,6 @@ import (
 	log "github.com/Sirupsen/logrus"
 )
 
-type (
-	Puller interface {
-		Pull(subscription string, pullrequest *pubsub.PullRequest) (*pubsub.PullResponse, error)
-		Acknowledge(subscription, ackId string) (*pubsub.Empty, error)
-		ModifyAckDeadline(subscription string, ackIds []string, ackDeadlineSeconds int64) (*pubsub.Empty, error)
-		Get(subscription string) (*pubsub.Subscription, error)
-	}
-
-	pubsubPuller struct {
-		subscriptionsService *pubsub.ProjectsSubscriptionsService
-	}
-)
-
-func (pp *pubsubPuller) Pull(subscription string, pullrequest *pubsub.PullRequest) (*pubsub.PullResponse, error) {
-	return pp.subscriptionsService.Pull(subscription, pullrequest).Do()
-}
-
-func (pp *pubsubPuller) Acknowledge(subscription, ackId string) (*pubsub.Empty, error) {
-	ackRequest := &pubsub.AcknowledgeRequest{
-		AckIds: []string{ackId},
-	}
-	return pp.subscriptionsService.Acknowledge(subscription, ackRequest).Do()
-}
-
-func (pp *pubsubPuller) ModifyAckDeadline(subscription string, ackIds []string, ackDeadlineSeconds int64) (*pubsub.Empty, error) {
-	req := &pubsub.ModifyAckDeadlineRequest{
-		AckDeadlineSeconds: ackDeadlineSeconds,
-		AckIds:             ackIds,
-	}
-	return pp.subscriptionsService.ModifyAckDeadline(subscription, req).Do()
-}
-
-func (pp *pubsubPuller) Get(subscription string) (*pubsub.Subscription, error) {
-	return pp.subscriptionsService.Get(subscription).Do()
-}
-
 type JobConfig struct {
 	Subscription string              `json:"subscription,omitempty"`
 	PullInterval int                 `json:"pull_interval,omitempty"`
